requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sentence-transformers==5.1.0
scikit-learn==1.7.1
pandas==2.1.4
numpy==1.25.2
python-dotenv==1.0.0
httpx==0.25.2

